requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.1
Pillow>=9.5.0
python-dotenv==1.0.0
pydantic>=2.4.0
aiofiles==23.2.1
